app/requirements.txt
requests==2.32.3
tenacity==8.5.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
boto3